fix(access): only admin can operate create update and delete category

+++ app/Http/Controllers/CategoryController.php
@@ -26,6 +26,8 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Category::class);
+
         $request->validate([
             'name' => ['required', 'string', 'min:2']
         ]);
@@ -57,6 +59,8 @@ class CategoryController extends Controller
      */
     public function update(Request $request, Category $category)
     {
+        $this->authorize('update', $category);
+
         $request->validate([
             'name' => ['required', 'string', 'min:2']
         ]);
@@ -76,6 +80,8 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        $this->authorize('update', $category);
+
         $category->delete();
 
         return response()->json([
